refactor: processMap moved to `TransactionPool`
