+++ src/chains/ethereum/ethereum/src/transaction-pool.ts
@@ -14,6 +14,7 @@ import {
 import { EthereumInternalOptions } from "@ganache/ethereum-options";
 import { Executables } from "./miner/executables";
 import { TypedTransaction } from "@ganache/ethereum-transaction";
+import { Ethereum } from "./api-types";
 
 /**
  * Checks if the `replacer` is eligible to replace the `replacee` transaction
@@ -452,6 +453,28 @@ export default class TransactionPool extends Emittery<{ drain: undefined }> {
     }
     return null;
   }
+
+  public processMap(map: Map<string, Heap<TypedTransaction>>) {
+    let res: Record<
+      string,
+      Record<string, Ethereum.Pool.Transaction<"private">>
+    > = {};
+    
+    for (let [_, { array, length }] of map) {
+      for (let i = 0; i < length; ++i) {
+        const transaction = array[i];
+        const from = transaction.from.toString();
+        if (res[from] === undefined) {
+          res[from] = {};
+        }
+        // The nonce keys are actual decimal numbers (as strings) and not
+        // hex literals (based on what geth returns).
+        const nonce = transaction.nonce.toBigInt().toString();
+        res[from][nonce] = transaction.toJSON() as Ethereum.Pool.Transaction<"private">;
+      }
+    }
+    return res;
+  };
 
   readonly drain = () => {
     // notify listeners (the blockchain, then the miner, eventually) that we
